requirements.txt
# requirements.txt — Movie Recommendation System
# Install with: pip install -r requirements.txt

streamlit>=1.32.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
